Add Comments to Code
Add comments to code to highlight the use of each function.
