+++ src/main.py
@@ -3,48 +3,11 @@ import time
 import matplotlib.pyplot as plt
 import networkx as nx
 
-from distances import *
-from search import *
 from target import *
 from tower import *
 from playable import *
 
-# def driver(graph_type, graph_size, tower_count=3,
-#           tower_pos="random", target_pos="random",
-#           tower_search="normal", target_movement="random"):
-#    graph_types = {"complete" + str(graph_size): nx.complete_graph(graph_size),
-#                   "erdos" + str(graph_size): nx.erdos_renyi_graph(graph_size, 0.15)
-#                   }
-#
-#    graph = graph_types[graph_type + str(graph_size)]
-#
-#    # Establish initial positioning
-#    location_start_time = time.time()
-#
-#    #
-#
-#    location_finish_time = time.time()
-#    location_elapsed_time = location_finish_time - location_start_time
-#
-#    # movement / searching
-#    # play the game
-#    plt.subplot(111)
-#    nx.draw_networkx(graph, with_labels=True, font_weight='bold')
-#    plt.show()
-#
-#    print("Time for Location: ", location_elapsed_time, end="")
-#    print(" Using Tower - ", tower_pos, " Target - ", target_pos)
-#    print(tabulate([
-#        ['Tower Position', tower_pos, location_elapsed_time],
-#        ['Target Position', target_pos, location_elapsed_time],
-#
-#        ['Tower Searching', tower_pos, location_elapsed_time, 7],
-#        ['Target Movement', target_pos, location_elapsed_time, 5]
-#
-#    ], headers=['', 'Type', 'Time', 'Turns']))
 
-
-# ToDo Fix Bug for a node with no possible movements (Disconnected graph)
 if __name__ == '__main__':
     found = False
     playable = False
@@ -54,14 +17,8 @@ if __name__ == '__main__':
     tower_type = [RandomTower(), HeuristicTower(), OptimalTower()]
     target_type = [RandomTarget(), HeuristicTarget(), OptimalTarget()]
 
-    #    graph_types = {"complete" + str(graph_size): nx.complete_graph(graph_size),
-    #                   "erdos" + str(graph_size): nx.erdos_renyi_graph(graph_size, 0.15)
-    #                   }
-    #
-    #    graph = graph_types[graph_type + str(graph_size)]
-
-    # Create the initial graph
-    graph = nx.erdos_renyi_graph(graph_size, 0.15)
+    graph_type = [nx.erdos_renyi_graph(graph_size, 0.15), nx.random_tree(graph_size)]
+    graph = graph_type[0]
 
     if playable:
         print("Playable Method")
@@ -71,7 +28,6 @@ if __name__ == '__main__':
         turn_number = 0
         visited = []
 
-        # ToDo Easy, Medium. Hard Towers
         tower = RandomTower()
         tower_location = tower.initial_position(graph, tower_count)
 
@@ -116,11 +72,6 @@ if __name__ == '__main__':
                             print(neighbour, end=' ')
                         print("")
                         value = input("Please enter a Node to move to from the above Possible Moves?\n")
-                    elif value == 'o':
-                        optimal = "Not Configured"
-                        print("The Optimal Move is ", optimal)
-                        while int(value) not in possible_moves:
-                            value = input("Please enter a Node to move to from the above Possible Moves?\n")
                     else:
                         try:
                             while int(value) not in possible_moves:
@@ -146,91 +97,91 @@ if __name__ == '__main__':
         print("-----------------\n")
 
         # Loop through all combinations of the Tower/Target
-        for tower in tower_type:
-            for target in target_type:
-                print("Target: ", type(tower).__name__)
-                print("Tower: ", type(target).__name__)
+        for graph in graph_type:
+            for tower in tower_type:
+                for target in target_type:
+                    print("Target: ", type(tower).__name__)
+                    print("Tower: ", type(target).__name__)
 
-                found = False
-                turn_number = 0
-                visited = []
-                target_movement_elapsed_time = 0
-                total_start_time = time.time()
+                    found = False
+                    turn_number = 0
+                    visited = []
+                    target_movement_elapsed_time = 0
+                    total_start_time = time.time()
 
-                # Establish initial positioning
-                tower_location_start_time = time.time()
-                tower_location = tower.initial_position(graph, tower_count)
-                tower_location_finish_time = time.time()
+                    # Establish initial positioning
+                    tower_location_start_time = time.time()
+                    tower_location = tower.initial_position(graph, tower_count)
+                    tower_location_finish_time = time.time()
 
-                tower_location_elapsed_time = tower_location_finish_time - tower_location_start_time
-                print("Initial Tower Location: ", tower_location)
+                    tower_location_elapsed_time = tower_location_finish_time - tower_location_start_time
+                    print("Initial Tower Location: ", tower_location)
 
-                target_location_start_time = time.time()
-                if type(target) == OptimalTarget:
-                    longest_path = find_optimal_node(graph, tower_location)[1][0]
-                    longest_path = list(map(int, longest_path))
+                    target_location_start_time = time.time()
+                    if type(target) == OptimalTarget:
+                        longest_path = find_optimal_node(graph, tower_location)[1][0]
+                        longest_path = list(map(int, longest_path))
 
-                    target_location = target.initial_location(graph, longest_path)
-                else:
-                    target_location = target.initial_location(graph, tower_location)
-
-                target_location_finish_time = time.time()
-                target_location_elapsed_time = target_location_finish_time - target_location_start_time
-
-                print("Initial Target Location: ", target_location)
-
-                # Add towers to visited so that Target cannot go there
-                for t in tower_location:
-                    visited.append(t)
-
-                distance_to_every_node = []
-                for t in tower_location:
-                    distance_to_every_node.append(populate_distance_table(graph, t))
-
-                # Each turn
-                while not found:
-                    found = is_found(graph, target_location, tower_location, visited, distance_to_every_node)
-
-                    if not found:
-                        possible_moves = []
-                        for i in graph.neighbors(target_location):
-                            if i not in visited:
-                                possible_moves.append(i)
-
-                        if len(possible_moves) == 0:
-                            print("No Moves Available!")
-                            found = True
-                        else:
-                            turn_number += 1
-                            visited.append(target_location)
-                            target_movement_start_time = time.time()
-                            if type(target) == OptimalTarget:
-                                target_location = target.next_move(longest_path, turn_number)
-                            elif type(target) == HeuristicTarget:
-                                target_location = target.heuristic_target_next_move(graph, tower_location, visited,
-                                                                                    distance_to_every_node, possible_moves)
-                            else:
-                                target_location = target.next_move(possible_moves, turn_number)
-                            target_movement_finish_time = time.time()
-
-                            target_movement_elapsed_time = target_movement_elapsed_time + \
-                                                           (target_movement_finish_time - target_movement_start_time)
-
+                        target_location = target.initial_location(graph, longest_path)
                     else:
-                        print("Found!")
+                        target_location = target.initial_location(graph, tower_location)
 
-                total_end_time = time.time()
-                total_elapsed_time = total_end_time - total_start_time
+                    target_location_finish_time = time.time()
+                    target_location_elapsed_time = target_location_finish_time - target_location_start_time
 
-                print("Target found in %d turns" % turn_number)
-                print("Tower Location : ", tower_location_elapsed_time, end='')
-                print("\tTarget Location : ", target_location_elapsed_time)
-                print("Target Movement : ", target_movement_elapsed_time)
-                print("Total Time : ", total_elapsed_time)
+                    print("Initial Target Location: ", target_location)
 
+                    # Add towers to visited so that Target cannot go there
+                    for t in tower_location:
+                        visited.append(t)
 
-                print("Target visited ", end='')
-                for t in visited[tower_count:]:
-                    print("%d -> " % t, end='')
-                print(target_location)
-                print("\n")
+                    distance_to_every_node = []
+                    for t in tower_location:
+                        distance_to_every_node.append(populate_distance_table(graph, t))
+
+                    # Each turn
+                    while not found:
+                        found = is_found(graph, target_location, tower_location, visited, distance_to_every_node)
+
+                        if not found:
+                            possible_moves = []
+                            for i in graph.neighbors(target_location):
+                                if i not in visited:
+                                    possible_moves.append(i)
+
+                            if len(possible_moves) == 0:
+                                print("No Moves Available!")
+                                found = True
+                            else:
+                                turn_number += 1
+                                visited.append(target_location)
+                                target_movement_start_time = time.time()
+                                if type(target) == OptimalTarget:
+                                    target_location = target.next_move(longest_path, turn_number)
+                                elif type(target) == HeuristicTarget:
+                                    target_location = target.heuristic_target_next_move(graph, tower_location, visited,
+                                                                                        distance_to_every_node, possible_moves)
+                                else:
+                                    target_location = target.next_move(possible_moves, turn_number)
+                                target_movement_finish_time = time.time()
+
+                                target_movement_elapsed_time = target_movement_elapsed_time + \
+                                                               (target_movement_finish_time - target_movement_start_time)
+
+                        else:
+                            print("Found!")
+
+                    total_end_time = time.time()
+                    total_elapsed_time = total_end_time - total_start_time
+
+                    print("Target found in %d turns" % turn_number)
+                    print("Tower Location : ", tower_location_elapsed_time, end='')
+                    print("\tTarget Location : ", target_location_elapsed_time)
+                    print("Target Movement : ", target_movement_elapsed_time)
+                    print("Total Time : ", total_elapsed_time)
+
+                    print("Target visited ", end='')
+                    for t in visited[tower_count:]:
+                        print("%d -> " % t, end='')
+                    print(target_location)
+                    print("\n")
