+++ src/tests/test_movement.py
@@ -2,7 +2,7 @@ import unittest
 import networkx as nx
 
 from treelib import *
-from src.tests.context import movement
+from src.tests.context import optimal
 
 
 class TestMovement(unittest.TestCase):
@@ -25,37 +25,37 @@ class TestMovement(unittest.TestCase):
         tower = [1, 2]
         visited = [3, 4]
         distances = [{0: 1, 1: 0, 2: 1, 3: 1}, {0: 1, 1: 1, 2: 0, 3: 1}]
-        self.assertEqual(True, movement.is_found(self.G, target, tower, visited, distances))
+        self.assertEqual(True, optimal.is_found(self.G, target, tower, visited, distances))
 
     def test_is_not_found(self):
         target = 0
         tower = [1, 2]
         visited = [3, 4]
         distances = [{0: -1, 1: 0, 2: -1, 3: -1}, {0: -1, 1: -1, 2: 0, 3: -1}]
-        self.assertEqual(False, movement.is_found(self.G, target, tower, visited, distances))
+        self.assertEqual(False, optimal.is_found(self.G, target, tower, visited, distances))
 
     def test_optimal_path_no_tower(self):
         target = 0
         towers = []
-        self.assertEqual((3, [['0', '1', '2'], ['0', '2', '1']]), movement.optimal_path(self.small, target, towers))
+        self.assertEqual((3, [['0', '1', '2'], ['0', '2', '1']]), optimal.optimal_path(self.small, target, towers))
 
     def test_optimal_path_tower(self):
         target = 0
         towers = [1]
-        self.assertEqual((2, [['0', '2']]), movement.optimal_path(self.small, target, towers))
+        self.assertEqual((2, [['0', '2']]), optimal.optimal_path(self.small, target, towers))
 
     def test_optimal_path_towers(self):
         target = 0
         towers = [1, 2]
-        self.assertEqual((1, [['0']]), movement.optimal_path(self.small, target, towers))
+        self.assertEqual((1, [['0']]), optimal.optimal_path(self.small, target, towers))
 
     def test_find_optimal_node_no_tower(self):
         towers = []
-        self.assertEqual((2, [['0', '1'], ['1', '0']]), movement.find_optimal_node(self.xsmall, towers))
+        self.assertEqual((2, [['0', '1'], ['1', '0']]), optimal.find_optimal_node(self.xsmall, towers))
 
     def test_find_optimal_node_tower(self):
         towers = [100]
-        self.assertEqual((5, [['1', '2', '3', '4', '5'], ['5', '4', '3', '2', '1']]), movement.find_optimal_node(self.test_graph, towers))
+        self.assertEqual((5, [['1', '2', '3', '4', '5'], ['5', '4', '3', '2', '1']]), optimal.find_optimal_node(self.test_graph, towers))
 
 
 if __name__ == '__main__':
